pyproject.toml
[project]
name = "spikebot-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "cryptography>=46.0.3",
    "discord-py>=2.6.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
